requirements.txt
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
alembic>=1.13
fastapi>=0.110
uvicorn[standard]>=0.27
python-jose[cryptography]>=3.3
python-multipart>=0.0.9
bcrypt>=4.1
pandas>=2.0
numpy>=1.26
scikit-learn>=1.4
